# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/presentation/{id}": {
      "get": {
        "operationId": "PresentationController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Presentation"
        ]
      },
      "patch": {
        "operationId": "PresentationController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePresentationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Presentation"
        ]
      },
      "delete": {
        "operationId": "PresentationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Presentation"
        ]
      }
    },
...
    "/submission": {
      "post": {
        "operationId": "SubmissionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Submission"
        ]
      },
      "get": {
        "operationId": "SubmissionController_findAll",
        "parameters": [
          {
            "name": "eventEditionId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "withoutPresentation",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "orderByProposedPresentation",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "showConfirmedOnly",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "mainAuthorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Submission"
        ]
      }
    },
    "/submission/{id}": {
      "get": {
        "operationId": "SubmissionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Submission"
        ]
      },
      "patch": {
        "operationId": "SubmissionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Submission"
        ]
      },
      "delete": {
        "operationId": "SubmissionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Submission"
        ]
      }
    },
    "/uploads": {
      "post": {
        "operationId": "UploadsController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Arquivo carregado com sucesso!"
          },
          "400": {
            "description": "Arquivo inválido (tamanho ou tipo)."
          }
        },
        "summary": "Faz upload de um arquivo PDF",
        "tags": [
          "Uploads"
        ]
      }
    },
    "/uploads/{filename}": {
      "delete": {
        "operationId": "UploadsController_deleteFile",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo deletado com sucesso!"
          },
          "404": {
            "description": "Arquivo não encontrado."
          }
        },
        "summary": "Deleta um arquivo pelo nome do arquivo",
        "tags": [
          "Uploads"
        ]
      },
      "get": {
        "operationId": "UploadsController_getFile",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Uploads"
        ]
      }
    },
...
    "/presentation-block": {
      "post": {
        "operationId": "PresentationBlockController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePresentationBlockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PresentationBlock"
        ]
      },
      "get": {
        "operationId": "PresentationBlockController_findAll",
        "parameters": [
          {
            "name": "eventEditionId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userIdOfPanelist",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PresentationBlock"
        ]
      }
    },
    "/presentation-block/event-edition/{eventEditionId}": {
      "get": {
        "operationId": "PresentationBlockController_findAllByEventEditionId",
        "parameters": [
          {
            "name": "eventEditionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PresentationBlock"
        ]
      }
    },
    "/presentation-block/{id}": {
      "get": {
        "operationId": "PresentationBlockController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PresentationBlock"
        ]
      },
      "patch": {
        "operationId": "PresentationBlockController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePresentationBlockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PresentationBlock"
        ]
      },
      "delete": {
        "operationId": "PresentationBlockController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PresentationBlock"
        ]
      }
    },
    "/presentation-block/{id}/presentations/swap": {
      "patch": {
        "operationId": "PresentationBlockController_swapPresentations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwapMultiplePresentationsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PresentationBlock"
        ]
      }
    },
...
  },
  "info": {
    "title": "WEPGCOMP API",
    "description": "Portal WEPGCOMP - API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdatePresentationDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateSubmissionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSubmissionDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePresentationBlockDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "General",
              "Presentation"
            ],
            "example": "Presentation"
          }
        },
        "required": [
          "type"
        ]
      },
      "UpdatePresentationBlockDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "General",
              "Presentation"
            ],
            "example": "Presentation"
          }
        }
      },
      "SwapMultiplePresentationsDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}